agents/vision-agent/requirements.txt
a2a-sdk[http-server]==0.3.2
google-genai
uvicorn
Pillow
pydantic
python-dotenv
